perf(av1): integer SAD in the motion search, a priced-once symbol table
Two more bit-exact costs off the RD search's inner loops:

- the motion search accumulated its SAD as one f64 conversion and dependent
  add per sample; a 32x32 block of 8-bit samples sums exactly in u32, and
  that sum converts exactly to f64, so the value is unchanged;
- `symbol_bits` called libm's log2 for every symbol of every candidate the
  search prices, but a CDF interval is a 15-bit width, so there are only
  32769 prices; they are now computed once into a table (each entry exactly
  the expression the call site used to evaluate).

instructions:u on the 24-frame 1280x720 sequence bench: 231.8G -> 209.0G
(-9.8%); output bytes unchanged (64931), BD-rate unchanged on all three gate
clips, gate wall 7.0/6.7/6.7 -> 6.3/6.0/6.2 s.

+++ crates/ec-av1/src/motion.rs
@@ -276,13 +276,18 @@ fn search_traced_from_step(
             reference, stride, ref_width, ref_height, x_q4, y_q4, block_w, block_h,
             &mut dst16, fctx,
         );
-        let sad: f64 = source
+        // lane-av1speed: accumulated as an integer (a block is at most
+        // 32x32 samples of at most 255 each, nowhere near `u32`, and an
+        // integer sum this small is exact in `f64` whatever order it is
+        // summed in) instead of one `f64` conversion and dependent add per
+        // sample. Same value, several times the throughput.
+        let sad: u32 = source
             .iter()
             .zip(dst16.iter())
-            .map(|(&a, &b)| f64::from((i32::from(a) - i32::from(b)).unsigned_abs()))
+            .map(|(&a, &b)| (i32::from(a) - i32::from(b)).unsigned_abs())
             .sum();
         let diff = (mv.0 - pred_mv.0, mv.1 - pred_mv.1);
-        sad + lambda * mv_bits(diff)
+        f64::from(sad) + lambda * mv_bits(diff)
     };
 
     let mut trace = Vec::new();
